+++ weave/mappers_publisher.py
@@ -1,4 +1,5 @@
 import typing
+import copy
 from weave import box, graph
 from weave.artifact_wandb import likely_commit_hash
 from weave.language_features.tagging import tag_store
@@ -52,17 +53,12 @@ class FunctionToPyFunction(mappers.Mapper):
 
 class ObjectToPyDict(mappers_python_def.ObjectToPyDict):
     def apply(self, obj):
-        # res = super().apply(obj)
-        # copy_obj = dataclasses.copy.copy(obj)
-        # for prop_name, prop_serializer in self._property_serializers.items():
-        #     if prop_name == "_name":
-        #         prop_name = "name"
-        #     if prop_serializer is not None:
-        #         try:
-        #             setattr(copy_obj, prop_name, res[prop_name])
-        #         except:
-        #             breakpoint()
-        # obj = copy_obj
+        res = super().apply(obj)
+        copy_obj = copy.copy(obj)
+        for prop_name, prop_serializer in self._property_serializers.items():
+            if prop_serializer is not None:
+                setattr(copy_obj, prop_name, res[prop_name])
+        obj = copy_obj
         return obj
